server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
pillow==10.0.1
httpx==0.25.0
python-dotenv==1.0.0
